Improve some code details, fix some bugs (#3161)
* refactor: reuse some methods

* refactor: remove unused field

* refactor: simplify some code

* fix: a potential bug

* style: add .gitattributes for code style

* fix: issue#2848 and issue#2988

+++ core/src/main/java/com/alibaba/fastjson2/writer/ObjectWriterImplDouble.java
@@ -22,7 +22,7 @@ final class ObjectWriterImplDouble
             return;
         }
 
-        double value = ((Double) object).doubleValue();
+        double value = (Double) object;
         if ((features & JSONWriter.Feature.WriteNonStringValueAsString.mask) != 0) {
             jsonWriter.writeString(value);
         } else {
@@ -51,7 +51,7 @@ final class ObjectWriterImplDouble
             return;
         }
 
-        double value = ((Double) object).doubleValue();
+        double value = (Double) object;
         if ((features & JSONWriter.Feature.WriteNonStringValueAsString.mask) != 0) {
             jsonWriter.writeString(value);
             return;
